+++ internal/datasource/os/file_helpers.go
@@ -1,0 +1,10 @@
+package os
+
+import (
+	"fmt"
+	"os"
+)
+
+func GetPermissions(fileMode os.FileMode) string {
+	return fmt.Sprintf("%#o", fileMode.Perm())
+}
