+++ internal/service/config/nginx_test.go
@@ -254,16 +254,3 @@ func TestNginx_Validate(t *testing.T) {
 		})
 	}
 }
-
-func TestGetPermissions(t *testing.T) {
-	file, err := os.CreateTemp(".", "get_permissions_test.txt")
-	defer os.Remove(file.Name())
-	assert.NoError(t, err)
-
-	info, err := os.Stat(file.Name())
-	assert.NoError(t, err)
-
-	permissions := getPermissions(info.Mode())
-
-	assert.Equal(t, "0600", permissions)
-}
